@@ -40,14 +40,14 @@
   },
   "browserslist": "> 0.25%, not dead",
   "dependencies": {
-    "@mitodl/course-search-utils": "^2.3.0",
+    "@mitodl/course-search-utils": "^2.3.3",
     "@sentry/browser": "^7.0.0",
     "array-flat-polyfill": "^1.0.1",
     "bootstrap": "^4.3.1",
     "casual": "^1.6.2",
-    "core-js": "^3.20.2",
+    "core-js": "^3.38.1",
     "es6-promise": "^4.2.8",
-    "fuse.js": "^6.6.2",
+    "fuse.js": "^7.0.0",
     "history": "^5.3.0",
     "isomorphic-fetch": "^3.0.0",
     "jquery": "^3.5.1",
@@ -61,25 +61,25 @@
     "posthog-js": "^1.225.0",
     "promise-polyfill": "^8.3.0",
     "ramda": "^0.30.0",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.0.0-beta.15",
-    "react-content-loader": "^5.1.2",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.14",
+    "react-bootstrap": "^2.0.0",
+    "react-content-loader": "^7.0.0",
+    "react-dom": "^18.2.14",
     "react-dotdotdot": "1.3.1",
-    "react-infinite-scroller": "^1.2.4",
+    "react-infinite-scroller": "^1.2.6",
     "sass": "1.77.8",
     "video.js": "^7.11.8",
     "videojs-transcript-ac": "^0.8.8",
     "videojs-youtube": "^2.6.1"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+    "@babel/preset-typescript": "^7.25.7",
     "@mitodl/ocw-to-hugo": "1.36.0",
     "@playwright/test": "^1.45.0",
     "@swc/core": "^1.3.14",
     "@swc/helpers": "^0.4.12",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@types/enzyme": "^3.10.11",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
+    "@testing-library/react": "15",
     "@types/inquirer": "^8.2.5",
     "@types/jest": "^27.4.0",
     "@types/jquery": "^3.5.11",
@@ -87,8 +87,9 @@
     "@types/lodash.uppercase": "^4.3.6",
     "@types/promise-polyfill": "^6.0.4",
     "@types/ramda": "^0.27.62",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-infinite-scroller": "^1.2.2",
+    "@types/react": "^18.2.14",
+    "@types/react-dom": "^18.2.14",
+    "@types/react-infinite-scroller": "^1.2.5",
     "@types/serve-handler": "^6.1.1",
     "@types/sinon": "^17.0.0",
     "@types/video.js": "^7.3.29",
@@ -105,10 +106,10 @@
     "dotenv-webpack": "^8.0.0",
     "envalid": "^8.0.0",
     "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^7.32.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-mitodl": "^2.0.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jest": "^28.0.0",
     "eslint-plugin-react": "^7.31.10",
     "eslint-plugin-react-hooks": "^4.6.0",
