@@ -1,7 +1,10 @@
 module.exports = api => {
   const config = {
     presets: ["@babel/preset-env", "@babel/preset-react"],
-    plugins: ["@babel/plugin-syntax-dynamic-import"]
+    plugins: [
+      "@babel/plugin-syntax-dynamic-import",
+      "@babel/plugin-proposal-private-property-in-object"
+    ]
   }
 
   if (api.env("production")) {
